fix(consent): extract adapters to decoupled module so registry.ts stays build-safe
Importing adapter constants from runtime/registry/*.ts pulled utils.ts -> nuxt/app into module evaluation, breaking nuxt-module-build prepare. Adapters are pure data, moved to a dedicated consent-adapters.ts using type-only imports of proxy types.
