+++ packages/script/src/runtime/registry/clarity.ts
@@ -1,4 +1,4 @@
-import type { ConsentAdapter, RegistryScriptInput } from '#nuxt-scripts/types'
+import type { RegistryScriptInput } from '#nuxt-scripts/types'
 import { useRegistryScript } from '../utils'
 import { ClarityOptions } from './schemas'
 
@@ -60,24 +60,4 @@ export function useScriptClarity<T extends ClarityApi>(
             window.clarity('consent', options.defaultConsent)
         },
   }), _options)
-}
-
-/**
- * GCMv2 -> Clarity consent adapter.
- * Clarity accepts a boolean cookie toggle; we project lossy from `analytics_storage`:
- * - `analytics_storage === 'granted'` -> `clarity('consent', true)`
- * - `analytics_storage === 'denied'`  -> `clarity('consent', false)`
- * - other GCM categories are ignored.
- */
-export const clarityConsentAdapter: ConsentAdapter<ClarityApi> = {
-  applyDefault(state, proxy) {
-    if (!state.analytics_storage)
-      return
-    proxy.clarity('consent', state.analytics_storage === 'granted')
-  },
-  applyUpdate(state, proxy) {
-    if (!state.analytics_storage)
-      return
-    proxy.clarity('consent', state.analytics_storage === 'granted')
-  },
 }
