+++ packages/script/src/runtime/registry/google-analytics.ts
@@ -1,4 +1,4 @@
-import type { ConsentAdapter, RegistryScriptInput } from '#nuxt-scripts/types'
+import type { RegistryScriptInput } from '#nuxt-scripts/types'
 import { useRegistryScript } from '#nuxt-scripts/utils'
 import { withQuery } from 'ufo'
 import { GoogleAnalyticsOptions } from './schemas'
@@ -110,19 +110,6 @@ export interface GoogleAnalyticsApi {
 export { GoogleAnalyticsOptions }
 
 export type GoogleAnalyticsInput = RegistryScriptInput<typeof GoogleAnalyticsOptions>
-
-/**
- * GCMv2 -> Google Analytics consent adapter.
- * GA consumes GCMv2 natively; this is a pass-through of the full state.
- */
-export const googleAnalyticsConsentAdapter: ConsentAdapter<GoogleAnalyticsApi> = {
-  applyDefault(state, proxy) {
-    proxy.gtag('consent', 'default', state as ConsentOptions)
-  },
-  applyUpdate(state, proxy) {
-    proxy.gtag('consent', 'update', state as ConsentOptions)
-  },
-}
 
 export function useScriptGoogleAnalytics<T extends GoogleAnalyticsApi>(_options?: GoogleAnalyticsInput & { onBeforeGtagStart?: (gtag: GTag) => void }) {
   return useRegistryScript<T, typeof GoogleAnalyticsOptions>(_options?.key || 'googleAnalytics', (options) => {
