+++ packages/script/src/runtime/registry/tiktok-pixel.ts
@@ -1,4 +1,4 @@
-import type { ConsentAdapter, RegistryScriptInput } from '#nuxt-scripts/types'
+import type { RegistryScriptInput } from '#nuxt-scripts/types'
 import { withQuery } from 'ufo'
 import { useRegistryScript } from '../utils'
 import { TikTokPixelOptions } from './schemas'
@@ -55,27 +55,6 @@ export interface TikTokPixelApi {
     /** Defer consent until an explicit grant/revoke. Available after the script loads. */
     holdConsent: () => void
   }
-}
-
-function applyTikTokConsent(state: { ad_storage?: 'granted' | 'denied' }, proxy: TikTokPixelApi) {
-  if (!state.ad_storage)
-    return
-  if (state.ad_storage === 'granted')
-    proxy.ttq.grantConsent()
-  else
-    proxy.ttq.revokeConsent()
-}
-
-/**
- * GCMv2 -> TikTok consent adapter.
- * TikTok only exposes a binary ad-storage toggle, so we project lossy:
- * - `ad_storage === 'granted'` -> `ttq.grantConsent()`
- * - `ad_storage === 'denied'`  -> `ttq.revokeConsent()`
- * - other GCM categories are ignored.
- */
-export const tiktokPixelConsentAdapter: ConsentAdapter<TikTokPixelApi> = {
-  applyDefault: applyTikTokConsent,
-  applyUpdate: applyTikTokConsent,
 }
 
 declare global {
